requirements.txt
aiogram==3.13.1
pydantic==1.10.13
numpy==2.0.2
opencv-python==4.10.0.84
matplotlib==3.9.2
python-dotenv
APScheduler
pillow
aiohttp
tqdm
flask
flask-admin
requests
python-jose[cryptography]
aiosqlite